mietrecht_agent/requirements.txt
Flask>=3.0.0
google-generativeai>=0.8.0
Flask-JWT-Extended>=4.6.0
werkzeug>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
stripe>=11.0.0
